Login, Logout resolve
