+++ src/main/java/com/culture/BAEUNDAY/config/SecurityConfig.java
@@ -1,12 +1,8 @@
 package com.culture.BAEUNDAY.config;
 
-import com.culture.BAEUNDAY.domain.enums.Role;
-import com.culture.BAEUNDAY.jwt.Custom.CustomLoginFilter;
-import com.culture.BAEUNDAY.jwt.Custom.CustomLogoutFilter;
+import com.culture.BAEUNDAY.domain.user.Role;
 import com.culture.BAEUNDAY.jwt.Custom.JWTCheckFilter;
 import com.culture.BAEUNDAY.jwt.JWTUtil;
-import com.culture.BAEUNDAY.repository.RefreshRepository;
-import com.culture.BAEUNDAY.service.RefreshTokenService;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
@@ -19,7 +15,6 @@ import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import org.springframework.security.web.authentication.logout.LogoutFilter;
 import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.CorsConfigurationSource;
 
@@ -30,19 +25,16 @@ import java.util.Collections;
 @RequiredArgsConstructor
 public class SecurityConfig {
 
-    private final AuthenticationConfiguration authenticationConfiguration;
     private final JWTUtil jwtUtil;
-    private final RefreshTokenService refreshTokenService;
-    private final RefreshRepository refreshRepository;
-
-    @Bean
-    public AuthenticationManager authenticationManager(AuthenticationConfiguration configuration) throws Exception {
-        return configuration.getAuthenticationManager();
-    }
 
     @Bean //사용자 비밀번호 암호화
     public BCryptPasswordEncoder passwordEncoder() {
         return new BCryptPasswordEncoder();
+    }
+
+    @Bean //AuthenticationManager 동적 생성을 위해 필요
+    public AuthenticationManager authenticationManager(AuthenticationConfiguration configuration) throws Exception {
+        return configuration.getAuthenticationManager();
     }
 
     @Bean
@@ -58,19 +50,16 @@ public class SecurityConfig {
         //경로별 인가 작업
         http.authorizeHttpRequests((auth) -> auth
                 .requestMatchers("/swagger-ui/**", "/v3/api-docs/**").permitAll()
-                .requestMatchers("/", "/login", "/logout", "/user/check-name", "/user/check-username", "/user/register", "/reissue").permitAll()
+                .requestMatchers("/", "/login", "/logout", "/user/login", "/user/logout", "/user/check-name", "/user/check-username", "/user/register", "/reissue").permitAll()
                 .requestMatchers("/test").hasRole(Role.USER.name())
                 .requestMatchers("/user/profile/**", "/heart/**", "/review/**", "/posts/**", "/reserve/**").hasAnyRole(Role.USER.name())
                 .anyRequest().authenticated());
 
-        http
-                .addFilterAt(new CustomLoginFilter(authenticationManager(authenticationConfiguration), jwtUtil, refreshTokenService), UsernamePasswordAuthenticationFilter.class);
 
         http
-                .addFilterBefore(new JWTCheckFilter(jwtUtil), CustomLoginFilter.class);
+            .addFilterBefore(new JWTCheckFilter(jwtUtil), UsernamePasswordAuthenticationFilter.class);
 
-        http
-                .addFilterBefore(new CustomLogoutFilter(jwtUtil, refreshRepository), LogoutFilter.class);
+
         //세션 설정 : STATELESS
         http
                 .sessionManagement((session) -> session
