--- src/main/java/com/culture/BAEUNDAY/domain/Heart.java
@@ -1,24 +1,0 @@
-package com.culture.BAEUNDAY.domain;
-
-import jakarta.persistence.*;
-
-@Entity
-@Table(name = "heart")
-public class Heart {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id")
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private User user;
-
-    /*
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id")
-    private Post post;
-    */
-
-}
