--- src/main/java/com/culture/BAEUNDAY/domain/Review.java
@@ -1,23 +1,0 @@
-package com.culture.BAEUNDAY.domain;
-
-import jakarta.persistence.*;
-
-@Entity
-@Table(name = "review")
-public class Review {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id")
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private User user;
-
-    @Column(nullable = false)
-    private String field;
-
-    private int star;
-
-}
