+++ src/main/java/com/culture/BAEUNDAY/domain/comment/Comment.java
@@ -1,0 +1,39 @@
+package com.culture.BAEUNDAY.domain.comment;
+
+import com.culture.BAEUNDAY.domain.reply.Reply;
+import com.culture.BAEUNDAY.domain.user.User;
+import jakarta.persistence.*;
+
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+
+@Entity
+@Table(name = "comment")
+public class Comment {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id")
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
+    private User user;
+
+    /*
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "post_id")
+    private Post post;
+    */
+
+    @Column(nullable = false)
+    private String content;
+
+    @Column(nullable = false)
+    private LocalDateTime createdDate;
+
+    @OneToMany(mappedBy = "comment")
+    private List<Reply> replies =  new ArrayList<>();
+
+}
