+++ src/main/java/com/culture/BAEUNDAY/domain/heart/Heart.java
@@ -1,0 +1,25 @@
+package com.culture.BAEUNDAY.domain.heart;
+
+import com.culture.BAEUNDAY.domain.user.User;
+import jakarta.persistence.*;
+
+@Entity
+@Table(name = "heart")
+public class Heart {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id")
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
+    private User user;
+
+    /*
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "post_id")
+    private Post post;
+    */
+
+}
