+++ src/main/java/com/culture/BAEUNDAY/domain/reply/Reply.java
@@ -1,0 +1,33 @@
+package com.culture.BAEUNDAY.domain.reply;
+
+import com.culture.BAEUNDAY.domain.comment.Comment;
+import com.culture.BAEUNDAY.domain.user.User;
+import jakarta.persistence.*;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Table(name = "reply")
+public class Reply {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id")
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
+    private User user;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "comment_id")
+    private Comment comment;
+
+    @Column(nullable = false)
+    private String content;
+
+    @Column(nullable = false)
+    private LocalDateTime createdDate;
+
+
+}
