+++ src/main/java/com/culture/BAEUNDAY/domain/review/Review.java
@@ -1,0 +1,24 @@
+package com.culture.BAEUNDAY.domain.review;
+
+import com.culture.BAEUNDAY.domain.user.User;
+import jakarta.persistence.*;
+
+@Entity
+@Table(name = "review")
+public class Review {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id")
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
+    private User user;
+
+    @Column(nullable = false)
+    private String field;
+
+    private int star;
+
+}
