+++ src/main/java/com/culture/BAEUNDAY/domain/user/DTO/request/CheckRequestDTO.java
@@ -1,0 +1,9 @@
+package com.culture.BAEUNDAY.domain.user.DTO.request;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record CheckRequestDTO(
+        @NotBlank(message = "필수 입력 항목입니다.")
+        String checkname
+) {
+}
