+++ src/main/java/com/culture/BAEUNDAY/domain/user/DTO/request/LoginDTO.java
@@ -1,0 +1,12 @@
+package com.culture.BAEUNDAY.domain.user.DTO.request;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record LoginDTO(
+        @NotBlank(message = "아이디는 필수 입력 항목입니다.")
+        String username,
+
+        @NotBlank(message = "비밀번호는 필수 입력 항목입니다.")
+        String password
+) {
+}
