+++ src/main/java/com/culture/BAEUNDAY/domain/user/DTO/response/CheckNameResponseDTO.java
@@ -1,0 +1,7 @@
+package com.culture.BAEUNDAY.domain.user.DTO.response;
+
+import lombok.Builder;
+
+@Builder
+public record CheckNameResponseDTO(String name, Integer successName, String message) {
+}
