+++ src/main/java/com/culture/BAEUNDAY/domain/user/DTO/response/CheckProfileResponseDTO.java
@@ -1,0 +1,11 @@
+package com.culture.BAEUNDAY.domain.user.DTO.response;
+
+import lombok.Builder;
+
+@Builder
+public record CheckProfileResponseDTO(
+        String name,
+        String profileImg,
+        String field
+) {
+}
