+++ src/main/java/com/culture/BAEUNDAY/domain/user/DTO/response/CheckUsernameResponseDTO.java
@@ -1,0 +1,7 @@
+package com.culture.BAEUNDAY.domain.user.DTO.response;
+
+import lombok.Builder;
+
+@Builder
+public record CheckUsernameResponseDTO(String username, Integer successUserName, String message) {
+}
