+++ src/main/java/com/culture/BAEUNDAY/domain/user/DTO/response/ErrorResult.java
@@ -1,0 +1,11 @@
+package com.culture.BAEUNDAY.domain.user.DTO.response;
+
+import lombok.Builder;
+
+@Builder
+public record ErrorResult(
+        Integer status,
+        String message,
+        String errorCode
+) {
+}
