+++ src/main/java/com/culture/BAEUNDAY/domain/user/Role.java
@@ -1,0 +1,17 @@
+package com.culture.BAEUNDAY.domain.user;
+
+import lombok.Getter;
+
+public enum Role {
+
+    ADMIN("ROLE_ADMIN"),
+    USER("ROLE_USER");
+
+    Role(String value) {
+        this.value = value;
+    }
+
+    @Getter
+    private String value;
+
+}
