+++ src/main/java/com/culture/BAEUNDAY/domain/user/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package com.culture.BAEUNDAY.domain.user.repository;
+
+import com.culture.BAEUNDAY.domain.user.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    boolean existsByName(String name);
+
+    boolean existsByUsername(String username);
+
+    User findByUsername(String username);
+}
