--- src/main/java/com/culture/BAEUNDAY/dto/request/CheckRequestDTO.java
@@ -1,9 +1,0 @@
-package com.culture.BAEUNDAY.dto.request;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record CheckRequestDTO(
-        @NotBlank(message = "필수 입력 항목입니다.")
-        String checkname
-) {
-}
