--- src/main/java/com/culture/BAEUNDAY/dto/request/LoginDTO.java
@@ -1,7 +1,0 @@
-package com.culture.BAEUNDAY.dto.request;
-
-public record LoginDTO(
-        String username,
-        String password
-) {
-}
