--- src/main/java/com/culture/BAEUNDAY/dto/response/CheckNameResponseDTO.java
@@ -1,7 +1,0 @@
-package com.culture.BAEUNDAY.dto.response;
-
-import lombok.Builder;
-
-@Builder
-public record CheckNameResponseDTO(String name, Integer successName, String message) {
-}
