--- src/main/java/com/culture/BAEUNDAY/dto/response/CheckProfileResponseDTO.java
@@ -1,11 +1,0 @@
-package com.culture.BAEUNDAY.dto.response;
-
-import lombok.Builder;
-
-@Builder
-public record CheckProfileResponseDTO(
-        String name,
-        String profileImg,
-        String field
-) {
-}
