--- src/main/java/com/culture/BAEUNDAY/dto/response/CheckUsernameResponseDTO.java
@@ -1,7 +1,0 @@
-package com.culture.BAEUNDAY.dto.response;
-
-import lombok.Builder;
-
-@Builder
-public record CheckUsernameResponseDTO(String username, Integer successUserName, String message) {
-}
