--- src/main/java/com/culture/BAEUNDAY/dto/response/ErrorResult.java
@@ -1,11 +1,0 @@
-package com.culture.BAEUNDAY.dto.response;
-
-import lombok.Builder;
-
-@Builder
-public record ErrorResult(
-        Integer status,
-        String message,
-        String errorCode
-) {
-}
