--- src/main/java/com/culture/BAEUNDAY/dto/response/RegisterResponseDTO.java
@@ -1,16 +1,0 @@
-package com.culture.BAEUNDAY.dto.response;
-
-import lombok.Builder;
-
-@Builder
-public record RegisterResponseDTO(
-        String message,
-        String name,
-        String username,
-        String profileImg,
-        Double manner
-) {
-
-
-
-}
