--- src/main/java/com/culture/BAEUNDAY/dto/response/UpdateProfileResponseDTO.java
@@ -1,14 +1,0 @@
-package com.culture.BAEUNDAY.dto.response;
-
-import lombok.Builder;
-
-@Builder
-public record UpdateProfileResponseDTO(
-        String message,
-        String name,
-        String field
-) {
-
-
-
-}
