--- src/main/java/com/culture/BAEUNDAY/repository/RefreshRepository.java
@@ -1,14 +1,0 @@
-package com.culture.BAEUNDAY.repository;
-
-import com.culture.BAEUNDAY.domain.Refresh;
-import jakarta.transaction.Transactional;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RefreshRepository extends JpaRepository<Refresh, Long> {
-
-    Boolean existsByRefresh(String refresh);
-
-    @Transactional
-    void deleteByRefresh(String refresh);
-
-}
