--- src/main/java/com/culture/BAEUNDAY/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package com.culture.BAEUNDAY.repository;
-
-import com.culture.BAEUNDAY.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    boolean existsByName(String name);
-
-    boolean existsByUsername(String username);
-
-    User findByUsername(String username);
-}
